test(parquet): move async reader tests to integration target
Keep private-API coverage in the source module and run public async reader tests through the Parquet integration target.
